+++ lib/src/builders/data/types/mixin_type_builder.dart
@@ -1,4 +1,3 @@
-import 'to_entity_type_builder.dart';
 import '../models/data_element.dart';
 import 'copy_with_type_builder.dart';
 import 'equals_type_builder.dart';
@@ -17,16 +16,12 @@ class MixinTypeBuilder implements TypeBuilder {
     final equals = EqualsTypeBuilder();
     final copyWith = CopyWithTypeBuilder();
     final toStringType = ToStringTypeBuilder();
-    final toEntity = ToEntityTypeBuilder();
 
     buffer.writeln('mixin $name {');
     buffer.writeln(fields.declaration(element: element));
     buffer.writeln(hashCode.declaration(element: element));
     buffer.writeln(equals.declaration(element: element));
     buffer.writeln(copyWith.declaration(element: element));
-    if (element.entity != null) {
-      buffer.writeln(toEntity.declaration(element: element));
-    }
     buffer.write(toStringType.declaration(element: element));
     buffer.writeln('}');