+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 1.0.0
+## 1.1.1
+- Fixed the entity feature from data builder
 
+## 1.1.0
+- Added Entity builder and entity features to Data builder
+
+## 1.0.0
 - Initial version